Create significant_days.py
This file analyzes a stock and several news articles pertaining to that stock and a significant day, which would be a movement of 3% or higher in this case. This would display that access news articles of any given day would help the prediction of that stock movement, if given live access to news articles.